% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{markets}
\alias{markets}
\title{Buda API Available Markets}
\usage{
markets(market_id = NULL)
}
\arguments{
\item{market_id}{A character string with valid market id (optional)}
}
\value{
A tibble with a row for each available market on the exchange, if a
market id is defined the tibble will have a single row with the information
of the selected market id.
}
\description{
\code{markets()} produces a \code{tibble} with the available markets information on Buda
cripto-exchange, the column names are kept the same as received from the
exchange.

Detailed descriptions about fields and Buda REST API can be found in
\link{https://api.buda.com}
}
\examples{
buda_markets <- markets()
BTC_CLP_market <- markets("BTC-CLP")

}
