% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{orders_book}
\alias{orders_book}
\title{Buda API Orders Book}
\usage{
orders_book(market_id)
}
\arguments{
\item{market_id}{A character string with valid market id (required)}
}
\value{
A tibble with the active orders for the selected market. Order type "ask" refers to sells and order type "bids" refers to buys.
}
\description{
Gets the list of active (open) orders on the selected market
}
\examples{
BTC_COP_orders_book <- orders_book("BTC-COP")
}
