% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{ticker}
\alias{ticker}
\title{Buda API Ticker of a Market}
\usage{
ticker(market_id)
}
\arguments{
\item{market_id}{A character string with valid market id (required)}
}
\value{
A tibble row with ticker for the specified market id.
}
\description{
A ticker allows to know the current status of a specific \code{market_id}. Details
regarding descriptions about fields and Buda REST API can be found in
\link{https://api.buda.com/#rest-api-llamadas-publicas-volumen-transado}
}
\examples{
BTC_COP_ticker <- ticker("BTC-COP")
}
