% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{tickers}
\alias{tickers}
\title{Buda API tickers}
\usage{
tickers()
}
\value{
A tibble with price information for each market id available on the exchange.
}
\description{
Shows the current state of the whole available markets on the exchange.
}
\examples{
all_tickers <- tickers()
}
