% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{trades}
\alias{trades}
\title{Buad API last trades}
\usage{
trades(market_id)
}
\arguments{
\item{market_id}{A character string with valid market id (required)}
}
\value{
A tibble with the last trades for the selected market.
}
\description{
Gets the list of the last trades on a specific market
}
\examples{
BTC_USDC_trades <- trades("BTC-USDC")
}
