% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_calls.R
\name{volume}
\alias{volume}
\title{Buda API Transactions Volume of a Market}
\usage{
volume(market_id)
}
\arguments{
\item{market_id}{A character string with valid market id (required)}
}
\value{
A tibble row with transaction volumes information for the specified market id.
}
\description{
Transactions volume for an specified \code{market_id}. Details regarding
descriptions about fields and Buda REST API can be found in
\link{https://api.buda.com/#rest-api-llamadas-publicas-volumen-transado}
}
\examples{
BTC_COP_volume <- volume("BTC-COP")

# Volumes for all markets available
market_ids <- markets() |> dplyr::pull(id)
volumes <- purrr::map_df(market_ids, volume)
}
